fix(#497): gate CI on clippy and clear the 79 accumulated warnings
CI ran fmt-check, build and test but never clippy, so warnings
accumulated unchecked: 46 in huck-syntax, 33 in huck-engine.

Mechanical bulk via `cargo clippy --fix` (let-chains, `?`, needless
`return`, redundant derefs). The rest by hand, with the judgment calls
recorded rather than silently applied:

- `Shell::report_error` was `pub` but takes `pub(crate) ErrorKind`
  (v358) — a private-in-public leak. Now `pub(crate)`; nothing outside
  the crate could name the type, so nothing breaks.
- `ErrorKind::BuiltinError` (v358) is never CONSTRUCTED, but it is
  documented and pinned by `ordinary_builtin_errors_always_continue`:
  `Continue` is the no-op verdict, so the ~360 emit-only builtin sites
  reach it by not calling `report_error` at all. KEPT with
  `#[allow(dead_code)]` + the reason, rather than deleting a measured
  default one day after it landed.
- `set_o_option_by_name` returned `Result<(), ()>`; now `bool`. One
  caller (`repl.rs`), which already used it as a boolean.
- `expand_part` (9 args) and `take_opt_value` (8) get narrow
  `#[allow(clippy::too_many_arguments)]` — grouping their parameters is
  a real refactor of the expansion core / flag parsing (#496), not a
  signature tweak.
- `Lexer::next` gets `#[allow(clippy::should_implement_trait)]`: lexing
  is fallible, a shape `Iterator` cannot express.
- Two doc lists (`lexer.rs`, `shell.rs`) were indented far enough to
  render as CODE BLOCKS in rustdoc; reflowed to real list continuation.
- Three orphaned doc comments described `old_cs`/`old_arith`/`map`,
  all deleted with the oracle (v266) or moved out (#458). Converted to
  plain comments so the history survives without documenting nothing.

Gate: `--all-targets`, so test code is covered too — a fifth of the
warnings were there.

Verification: clippy clean workspace-wide under `-D warnings`; lib tests
483 (syntax) + 1994 (engine), both at baseline; diff-check sweep
274/274.

Closes #497

+++ crates/huck-engine/src/jobs.rs
@@ -463,7 +463,7 @@ pub fn reap_owned_once(shell: &mut crate::shell_state::Shell) -> bool {
     // coproc list) as it reaps. Coproc pids are tracked on the Shell, not the job
     // table, so they are unioned in here.
     let mut targets: Vec<i32> = shell.jobs.owned_pids().to_vec();
-    targets.extend(shell.coprocs.iter().map(|c| c.pid as i32));
+    targets.extend(shell.coprocs.iter().map(|c| c.pid));
     targets.sort_unstable();
     targets.dedup();
